fix(localization): make inflection validation deterministic
